Add working skeleton for canStartBuild method
The problem is that the BuildRequests launch so fast that any method of checking
CPU, load average, docker container count etc. is doomed to fail, unless hacky
random guaranteens are added. Even then results ("how many builds will run at any
given time") are unsure. This affects the initial launch of the build, because
after build has progressed far enough load average numbers actually give useful
numbers. The proper solution would be to create an Interlock that operates on the
level of a group of workers. Like this, but updated to recent buildbot:

https://lists.buildbot.net/pipermail/commits/2013-June.txt

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>