Provision msibuilder the same way as buildbot Windows workers
Lots of changes had been made to provisioning of Windows workers in
openvpn-buildbot. Make sure msibuilder's provisioning is using the same scripts
and procedures to allowing fixing issues found in one to be fixed trivially in
the other without further, extensive debugging and testing. The main change is
to move the Powershell provisioning scripts under "scripts", so that that
directory can be easily synced between openvpn-vagrant and openvpn-buildbot.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>